requirements.txt
fastapi
uvicorn
langgraph
langchain-google-vertexai
pinecone-client
pydantic-settings
python-dotenv
googlemaps